[Refactor] Notification.Name, UserInfoKeys 타입을 Plane에서만 사용되도록 변경
