--- DrawingApp/DrawingApp/Models/Constants/UserInfoKeys.swift
@@ -1,15 +1,0 @@
-//
-//  UserInfoKeys.swift
-//  DrawingApp
-//
-//  Created by 김상혁 on 2022/03/17.
-//
-
-import Foundation
-
-enum UserInfoKeys {
-    static let newRectangle = "NewRectangle"
-    static let updatedSelectedShape = "UpdatedSelectedShape"
-    static let tappedShape = "TappedShape"
-    static let selectedShape = "SelectedShape"
-}
