+++ DrawingApp/DrawingApp/Models/Plane.swift
@@ -23,14 +23,14 @@ class Plane {
     func addRectangle(bound: (x: Double, y: Double), by factoryType: RandomShapeFactory.Type) {
         let newRectangle = factoryType.createRandomRectangle(xBound: bound.x, yBound: bound.y)
         shapes.append(newRectangle)
-        NotificationCenter.default.post(name: .newShapeDidCreate,
+        NotificationCenter.default.post(name: Plane.EventName.newShapeDidCreate,
                                         object: self,
                                         userInfo: [UserInfoKeys.newRectangle: newRectangle])
     }
     
     func updateSelected(shape: BasicShape) {
         selected = shape
-        NotificationCenter.default.post(name: .selectedShapeDidChange,
+        NotificationCenter.default.post(name: Plane.EventName.selectedShapeDidChange,
                                         object: self,
                                         userInfo: [UserInfoKeys.updatedSelectedShape: shape])
     }
@@ -53,14 +53,14 @@ class Plane {
     func searchTopShape(on coordinate: (x: Double, y: Double)) {
         for shape in shapes.reversed() {
             if isShapeExist(on: coordinate, target: shape) {
-                NotificationCenter.default.post(name: .shapeDidSelect,
+                NotificationCenter.default.post(name: Plane.EventName.shapeDidSelect,
                                                 object: self,
                                                 userInfo: [UserInfoKeys.tappedShape: shape])
                 return
             }
         }
         
-        NotificationCenter.default.post(name: .emptyViewDidSelect,
+        NotificationCenter.default.post(name: Plane.EventName.emptyViewDidSelect,
                                         object: self)
     }
     
@@ -68,7 +68,7 @@ class Plane {
         guard let selectedShape = selected as? BasicShape & Colorable else { return }
         let randomColor = RandomAttributeFactory.generateRandomColor()
         selectedShape.changeBackgroundColor(by: randomColor)
-        NotificationCenter.default.post(name: .selectedShapeDidUpdateBackgroundColor,
+        NotificationCenter.default.post(name: Plane.EventName.selectedShapeDidUpdateBackgroundColor,
                                         object: self,
                                         userInfo: [UserInfoKeys.selectedShape: selectedShape])
     }
@@ -76,7 +76,7 @@ class Plane {
     func upAlphaValue() {
         guard let selectedShape = selected as? BasicShape & Alphable else { return }
         selectedShape.upAlphaLevel()
-        NotificationCenter.default.post(name: .selectedShapeDidUpdateAlpha,
+        NotificationCenter.default.post(name: Plane.EventName.selectedShapeDidUpdateAlpha,
                                         object: self,
                                         userInfo: [UserInfoKeys.selectedShape: selectedShape])
     }
@@ -84,8 +84,30 @@ class Plane {
     func downAlphaValue() {
         guard let selectedShape = selected as? BasicShape & Alphable else { return }
         selectedShape.downAlphaLevel()
-        NotificationCenter.default.post(name: .selectedShapeDidUpdateAlpha,
+        NotificationCenter.default.post(name: Plane.EventName.selectedShapeDidUpdateAlpha,
                                         object: self,
                                         userInfo: [UserInfoKeys.selectedShape: selectedShape])
     }
 }
+
+//MARK: Notifications
+
+extension Plane {
+    enum UserInfoKeys {
+        static let newRectangle = "NewRectangle"
+        static let updatedSelectedShape = "UpdatedSelectedShape"
+        static let tappedShape = "TappedShape"
+        static let selectedShape = "SelectedShape"
+    }
+    
+    enum EventName {
+        static let newShapeDidCreate = NSNotification.Name("NewShapeDidCreateNotification")
+        static let shapeDidSelect = NSNotification.Name("ShapeDidSelectNotification")
+        static let emptyViewDidSelect = NSNotification.Name("EmptyViewDidSelectNotification")
+
+        static let selectedShapeDidChange = NSNotification.Name("SelectedShapeDidChangeNotification")
+        static let selectedShapeDidUpdateBackgroundColor = NSNotification.Name("SelectedShapeDidUpdateBackgroundColorNotification")
+        static let selectedShapeDidUpdateAlpha = NSNotification.Name("SelectedShapeDidUpdateAlphaNotification")
+    }
+}
+
